requirements.txt
django==2.0.9
factory-boy==2.11.1
